Fix typecheck for mobile after AudiusBackend ts migration
