+++ packages/web/src/services/AudiusBackend.ts
@@ -66,10 +66,40 @@ import {
 } from './audius-backend/eagerLoadUtils'
 import { monitoringCallbacks } from './serviceMonitoring'
 
+type DisplayEncoding = 'utf8' | 'hex'
+type PhantomEvent = 'disconnect' | 'connect' | 'accountChanged'
+type PhantomRequestMethod =
+  | 'connect'
+  | 'disconnect'
+  | 'signTransaction'
+  | 'signAllTransactions'
+  | 'signMessage'
+
+interface ConnectOpts {
+  onlyIfTrusted: boolean
+}
+export interface PhantomProvider {
+  publicKey: PublicKey | null
+  isConnected: boolean | null
+  isPhantom: boolean
+  signTransaction: (transaction: Transaction) => Promise<Transaction>
+  signAndSendTransaction: (transaction: Transaction) => Promise<Transaction>
+  signAllTransactions: (transactions: Transaction[]) => Promise<Transaction[]>
+  signMessage: (
+    message: Uint8Array | string,
+    display?: DisplayEncoding
+  ) => Promise<any>
+  connect: (opts?: Partial<ConnectOpts>) => Promise<{ publicKey: PublicKey }>
+  disconnect: () => Promise<void>
+  on: (event: PhantomEvent, handler: (args: any) => void) => void
+  request: (method: PhantomRequestMethod, params: any) => Promise<unknown>
+}
 declare global {
   interface Window {
     web3Loaded: boolean
     phantom: any
+    solana: PhantomProvider
+    web3: any
   }
 }
 
@@ -377,7 +407,7 @@ class AudiusBackend {
   static getTrackImages(track: TrackMetadata) {
     const coverArtSizes: CoverArtSizes = {}
     if (!track.cover_art_sizes && !track.cover_art) {
-      coverArtSizes[DefaultSizes.OVERRIDE] = placeholderCoverArt
+      coverArtSizes[DefaultSizes.OVERRIDE] = placeholderCoverArt as string
     }
 
     return {
@@ -405,7 +435,7 @@ class AudiusBackend {
     }
 
     if (!collection.cover_art_sizes && !collection.cover_art) {
-      coverArtSizes[DefaultSizes.OVERRIDE] = placeholderCoverArt
+      coverArtSizes[DefaultSizes.OVERRIDE] = placeholderCoverArt as string
     }
 
     return {
@@ -421,11 +451,12 @@ class AudiusBackend {
     // Images are fetched on demand async w/ the `useUserProfilePicture`/`useUserCoverPhoto` and
     // transitioned in w/ the dynamicImageComponent
     if (!user.profile_picture_sizes && !user.profile_picture) {
-      profilePictureSizes[DefaultSizes.OVERRIDE] = placeholderProfilePicture
+      profilePictureSizes[DefaultSizes.OVERRIDE] =
+        placeholderProfilePicture as string
     }
 
     if (!user.cover_photo_sizes && !user.cover_photo) {
-      coverPhotoSizes[DefaultSizes.OVERRIDE] = imageCoverPhotoBlank
+      coverPhotoSizes[DefaultSizes.OVERRIDE] = imageCoverPhotoBlank as string
     }
 
     return {
@@ -494,7 +525,7 @@ class AudiusBackend {
     SolanaUtils = libs.SolanaUtils
 
     // initialize libs
-    let libsError = null
+    let libsError: Nullable<string> = null
     const { web3Config } = await AudiusBackend.getWeb3Config()
     const { ethWeb3Config } = AudiusBackend.getEthWeb3Config()
     const { solanaWeb3Config } = AudiusBackend.getSolanaWeb3Config()
@@ -1466,7 +1497,7 @@ class AudiusBackend {
 
       if (error) return { playlistId, error }
 
-      const updatePromises = []
+      const updatePromises: Promise<any>[] = []
 
       // If this playlist is being created from an existing cover art, use it.
       if (metadata.cover_art_sizes) {
@@ -1515,7 +1546,7 @@ class AudiusBackend {
 
     try {
       let blockHash, blockNumber
-      const promises = []
+      const promises: Promise<any>[] = []
       if (playlistName) {
         promises.push(
           audiusLibs.Playlist.updatePlaylistName(playlistId, playlistName)
