+++ packages/web/src/services/web3-modal/index.ts
@@ -1,4 +1,3 @@
-import { PublicKey, Transaction } from '@solana/web3.js'
 import Web3Modal, { IProviderOptions } from 'web3modal'
 
 import phantomIconPurpleSvg from 'assets/img/phantom-icon-purple.svg'
@@ -12,42 +11,6 @@ const WEB3_NETWORK_ID = parseInt(process.env.REACT_APP_ETH_NETWORK_ID || '')
 const ETH_PROVIDER_URLS = (process.env.REACT_APP_ETH_PROVIDER_URL || '').split(
   ','
 )
-
-type DisplayEncoding = 'utf8' | 'hex'
-type PhantomEvent = 'disconnect' | 'connect' | 'accountChanged'
-type PhantomRequestMethod =
-  | 'connect'
-  | 'disconnect'
-  | 'signTransaction'
-  | 'signAllTransactions'
-  | 'signMessage'
-
-interface ConnectOpts {
-  onlyIfTrusted: boolean
-}
-
-export interface PhantomProvider {
-  publicKey: PublicKey | null
-  isConnected: boolean | null
-  isPhantom: boolean
-  signTransaction: (transaction: Transaction) => Promise<Transaction>
-  signAndSendTransaction: (transaction: Transaction) => Promise<Transaction>
-  signAllTransactions: (transactions: Transaction[]) => Promise<Transaction[]>
-  signMessage: (
-    message: Uint8Array | string,
-    display?: DisplayEncoding
-  ) => Promise<any>
-  connect: (opts?: Partial<ConnectOpts>) => Promise<{ publicKey: PublicKey }>
-  disconnect: () => Promise<void>
-  on: (event: PhantomEvent, handler: (args: any) => void) => void
-  request: (method: PhantomRequestMethod, params: any) => Promise<unknown>
-}
-
-declare global {
-  interface Window {
-    solana: PhantomProvider
-  }
-}
 
 type Config = {
   isBitSkiEnabled: boolean
